follow vault control
