+++ features/follow/common/FollowButton.tsx
@@ -1,14 +1,11 @@
 import { Icon } from '@makerdao/dai-ui-icons'
-import { AppSpinner } from 'helpers/AppSpinner'
 import { useTranslation } from 'next-i18next'
-import { Box, Button } from 'theme-ui'
-
-export type FollowButtonState = 'follow' | 'unfollow'
+import { Button, Flex, Spinner } from 'theme-ui'
 
 interface FollowButtonProps {
   isProcessing: boolean
   // isRetry: boolean
-  state: FollowButtonState
+  isFollowing: boolean
   buttonClickHandler: () => void
 }
 
@@ -16,36 +13,15 @@ export function FollowButton(props: FollowButtonProps) {
   const { t } = useTranslation()
 
   return (
-    <Box>
-      {props.isProcessing ? (
-        <AppSpinner
-          sx={{
-            position: 'relative',
-            left: '14px',
-            top: '25px',
-            width: '14px',
-            height: '14px',
-            transition: '0.2s',
-          }}
-        />
-      ) : (
-        <Icon
-          name={props.state === 'follow' ? 'star_empty' : 'star'}
-          size="15px"
-          sx={{
-            position: 'relative',
-            left: '28px',
-            top: '1px',
-            width: '14px',
-            height: '14px',
-            transition: '0.2s',
-            color: '#878BFC',
-          }}
-        />
-      )}
-      <Button variant="menuButton" disabled={props.isProcessing} onClick={props.buttonClickHandler}>
-        {props.state === 'follow' ? t('follow') : t('unfollow')}
-      </Button>
-    </Box>
+    <Button variant="menuButton" disabled={props.isProcessing} onClick={props.buttonClickHandler}>
+      <Flex sx={{ justifyContent: 'space-between', gap: '8px', alignItems: 'center' }}>
+        {props.isProcessing ? (
+          <Spinner size="10px" />
+        ) : (
+          <Icon name={props.isFollowing ? 'star_empty' : 'star'} size="15px" />
+        )}
+        {props.isFollowing ? t('follow') : t('unfollow')}
+      </Flex>
+    </Button>
   )
 }
